+++ packages/ricos-editor/web/src/toolbars/utils/biUtils.ts
@@ -1,4 +1,13 @@
-import { Version, Helpers, ToolbarType } from 'wix-rich-content-common';
+import {
+  Version,
+  Helpers,
+  ToolbarType,
+  ANCHOR_CATEGORY,
+  WEB_ADDRESS_CATEGORY,
+  ADD_PLUGIN_LINK_BI,
+  OnPluginAction,
+  OnAddPluginLink,
+} from 'wix-rich-content-common';
 
 export const getBiFunctions = (helpers: Helpers, contentId?: string) => {
   const onInlineToolbarOpen = toolbarType =>
@@ -7,6 +16,7 @@ export const getBiFunctions = (helpers: Helpers, contentId?: string) => {
       version: Version.currentVersion,
       contentId,
     });
+
   const onToolbarButtonClick = (
     name: string,
     toolbarType: ToolbarType,
@@ -32,5 +42,22 @@ export const getBiFunctions = (helpers: Helpers, contentId?: string) => {
       );
     }
   };
-  return { onInlineToolbarOpen, onToolbarButtonClick };
+
+  const onAddPluginLink: OnAddPluginLink = (data, plugin_id) => {
+    const { url, anchor, target, rel } = data;
+    const params = anchor
+      ? { anchor, category: ANCHOR_CATEGORY }
+      : {
+          link: url,
+          rel,
+          newTab: target === '_blank',
+          category: WEB_ADDRESS_CATEGORY,
+        };
+    onPluginAction(ADD_PLUGIN_LINK_BI, { plugin_id, params });
+  };
+
+  const onPluginAction: OnPluginAction = (eventName, params) =>
+    helpers.onPluginAction?.(eventName, { ...params, version: Version.currentVersion });
+
+  return { onInlineToolbarOpen, onToolbarButtonClick, onAddPluginLink };
 };